movie/serie btn updates
