+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import { useCallback, useRef } from "react";
 // import InfiniteScroll from "react-infinite-scroll-component";
-import FilterBar from "./components/FilterBar";
+// import FilterBar from "./components/FilterBar";
 import FilterBadge from "./components/FilterBadge";
 import "./style/App.css";
 import MainResearch from "./components/MainResearch";
@@ -27,46 +27,35 @@ function App() {
   );
 
   // Choisir le endpoint films
-  const handleClickMovies = () => {
-    searchContext.setPageNumber(1);
-    searchContext.setTypeVideo("movie");
-  };
+  // const handleClickMovies = () => {
+  //   searchContext.setPageNumber(1);
+  //   searchContext.setTypeVideo("movie");
+  // };
   // Choisir le endpoint series
-  const handleClickSeries = () => {
-    searchContext.setPageNumber(1);
-    searchContext.setTypeVideo("tv");
-  };
+  // const handleClickSeries = () => {
+  //   searchContext.setPageNumber(1);
+  //   searchContext.setTypeVideo("tv");
+  // };
   // Afficher ou faire disparaitre la section filtres
-  const handleClickFilters = () => {
-    searchContext.setFilters(!searchContext.filters);
-  };
+  // const handleClickFilters = () => {
+  //   searchContext.setFilters(!searchContext.filters);
+  // };
 
   return (
-    <>
-      <div className="container">
-        <FilterBar
-          handleClickMovies={handleClickMovies}
-          handleClickSeries={handleClickSeries}
-          handleClickFilters={handleClickFilters}
-          typeVideo={searchContext.typeVideo}
-        />
-      </div>
-
-      <div className="main-area pos-r">
-        <FilterBadge
-          setGenres={searchContext.setGenres}
-          handleClickFilters={handleClickFilters}
-          setPageNumber={searchContext.setPageNumber}
-          typeVideo={searchContext.typeVideo}
-        />
-        {searchContext.textFound !== "" && (
-          <MainResearch lastMovieElementRef={lastMovieElementRef} />
-        )}
-        {searchContext.textFound === "" && (
-          <PopularVideos typeVideo={searchContext.typeVideo} />
-        )}
-      </div>
-    </>
+    <div className="main-area pos-r">
+      <FilterBadge
+        setGenres={searchContext.setGenres}
+        // handleClickFilters={handleClickFilters}
+        setPageNumber={searchContext.setPageNumber}
+        typeVideo={searchContext.typeVideo}
+      />
+      {searchContext.textFound !== "" && (
+        <MainResearch lastMovieElementRef={lastMovieElementRef} />
+      )}
+      {searchContext.textFound === "" && (
+        <PopularVideos typeVideo={searchContext.typeVideo} />
+      )}
+    </div>
   );
 }
