+++ frontend/src/components/FilterBar.jsx
@@ -1,23 +1,34 @@
-import PropTypes from "prop-types";
+// import PropTypes from "prop-types";
+import { UseSearch } from "../contexts/SearchContext";
 
-function FilterBar({
-  handleClickMovies,
-  handleClickSeries,
-  handleClickFilters,
-}) {
+function FilterBar() {
+  // {
+  // handleClickMovies,
+  // handleClickSeries,
+  // handleClickFilters,
+  // }
+  const searchContext = UseSearch();
   return (
     <div className="filter-bar">
       <div className="switch-button big-switch-button">
-        <button onClick={(event) => handleClickMovies(event)} type="submit">
+        <button
+          onClick={(event) => searchContext.handleClickMovies(event)}
+          className={`${searchContext.typeVideo === "movie" && "active"}`}
+          type="button"
+        >
           Films
         </button>
-        <button onClick={(event) => handleClickSeries(event)} type="submit">
+        <button
+          onClick={(event) => searchContext.handleClickSeries(event)}
+          className={`${searchContext.typeVideo === "tv" && "active"}`}
+          type="button"
+        >
           Séries
         </button>
       </div>
       <div>
         <button
-          onClick={(event) => handleClickFilters(event)}
+          onClick={(event) => searchContext.handleClickFilters(event)}
           type="submit"
           className="filter-button"
         >
@@ -33,8 +44,8 @@ function FilterBar({
 }
 
 FilterBar.propTypes = {
-  handleClickMovies: PropTypes.func.isRequired,
-  handleClickSeries: PropTypes.func.isRequired,
-  handleClickFilters: PropTypes.func.isRequired,
+  // handleClickMovies: PropTypes.func.isRequired,
+  // handleClickSeries: PropTypes.func.isRequired,
+  // handleClickFilters: PropTypes.func.isRequired,
 };
 export default FilterBar;
